+++ packages/core/src/analytics/analytics-tracker-window.ts
@@ -1,7 +1,0 @@
-export interface AnalyticsTracker {
-    track(step: string, data: any): void;
-}
-
-export default interface AnalyticsTrackerWindow extends Window {
-    analytics: AnalyticsTracker;
-}