+++ packages/core/src/analytics/is-analytics-step-tracker-window.ts
@@ -1,5 +1,0 @@
-import AnalyticsTrackerWindow from './analytics-tracker-window';
-
-export function isAnalyticsTrackerWindow(window: Window): window is AnalyticsTrackerWindow {
-    return Boolean((window as AnalyticsTrackerWindow).analytics);
-}