+++ packages/payment-integration-api/src/address/legacy-address.ts
@@ -1,0 +1,12 @@
+export default interface LegacyAddress {
+    email: string;
+    first_name: string;
+    last_name: string;
+    phone_number: string;
+    address_line_1: string;
+    address_line_2: string;
+    city: string;
+    state: string;
+    country_code: string;
+    postal_code: string;
+}